feat: use exchangeRateHypothetical
