--- pkg/fork-tests/tests/20221130-midas-linear-pool/output/test.json
@@ -1,3 +1,0 @@
-{
-  "MidasLinearPoolFactory": "0x2b73D2725101E03860A64dfD2Df27d031aBE4076"
-}
