--- pkg/fork-tests/tests/20221130-midas-linear-pool/readme.md
@@ -1,5 +1,0 @@
-# 2022-11-30 - Midas Linear Pool
-
-Deployment of the `MidasLinearPoolFactory`, for Linear Pools with a cToken. This deployment includes:
-
-## Useful Files
