+++ pkg/fork-tests/tests/20230310-midas-rebalanced-linear-pool/output/test.json
@@ -1,0 +1,3 @@
+{
+  "MidasLinearPoolFactory": "0xfD14ac4867954173552dFB72be1321Ce1AC35834"
+}
