+++ pkg/fork-tests/tests/20230310-midas-rebalanced-linear-pool/test/test-4-decimals-polygon.fork.ts
@@ -1,0 +1,346 @@
+import hre from 'hardhat';
+import { expect } from 'chai';
+import { Contract } from 'ethers';
+import { setCode } from '@nomicfoundation/hardhat-network-helpers';
+import * as expectEvent from '@orbcollective/shared-dependencies/expectEvent';
+
+import { bn, fp, FP_ONE } from '@orbcollective/shared-dependencies/numbers';
+import {
+  MAX_UINT256,
+  getExternalPackageDeployedAt,
+  getExternalPackageArtifact,
+} from '@orbcollective/shared-dependencies';
+import { SignerWithAddress } from '@nomiclabs/hardhat-ethers/dist/src/signer-with-address';
+
+import { SwapKind } from '@balancer-labs/balancer-js';
+
+import { describeForkTest } from '../../../src/forkTests';
+import { impersonate, getForkedNetwork, Task, TaskMode, getSigners } from '../../../src';
+
+describeForkTest('polygon - MidasLinearPoolFactory - 4 decimals', 'polygon', 40164926, function () {
+  let owner: SignerWithAddress, holder: SignerWithAddress, other: SignerWithAddress;
+  let factory: Contract, vault: Contract, brz: Contract;
+  let rebalancer: Contract;
+
+  let task: Task;
+
+  //polygon
+  const cBRZ = '0x2e4659b451C3ba2E72D79aAf267cFc09BCCc9d7c';
+  const BRZ = '0x491a4eB4f1FC3BfF8E1d2FC856a6A46663aD556f';
+
+  //const BRZ = '0x71be881e9C5d4465B3FfF61e89c6f3651E69B5bb';
+  //const cBRZ = '0x9c42b10a32d7dfE1a978b74dcD07Eb1Fadea2B86';
+
+  const BRZ_SCALING = bn(1e14); // BRZ has 4 decimals, so its scaling factor is 1e14
+
+  //polygon
+  const BRZ_HOLDER = '0xB90B2050C955cd899b9BC8B5C743c25770EBc8AA';
+
+  //const BRZ_HOLDER = '0x6b02568149b01f88152c9a4f1853ec9c59922620';
+
+  const SWAP_FEE_PERCENTAGE = fp(0.01); // 1%
+
+  // The targets are set using 18 decimals, even if the token has fewer (as is the case for BRZ);
+  const INITIAL_UPPER_TARGET = fp(1e6);
+
+  // The initial midpoint (upper target / 2) must be between the final lower and upper targets
+  const FINAL_LOWER_TARGET = fp(0.2e6);
+  const FINAL_UPPER_TARGET = fp(5e6);
+
+  const PROTOCOL_ID = 0;
+
+  let pool: Contract;
+  let poolId: string;
+
+  before('run task', async () => {
+    task = new Task('20230310-midas-rebalanced-linear-pool', TaskMode.TEST, getForkedNetwork(hre));
+    await task.run({ force: true });
+    factory = await task.deployedInstance('MidasLinearPoolFactory');
+  });
+
+  before('load signers', async () => {
+    [, owner, other] = await getSigners();
+
+    holder = await impersonate(BRZ_HOLDER, fp(100));
+  });
+
+  before('setup contracts', async () => {
+    vault = await new Task('20210418-vault', TaskMode.READ_ONLY, getForkedNetwork(hre)).deployedInstance('Vault');
+
+    brz = await task.instanceAt('IERC20', BRZ);
+    await brz.connect(holder).approve(vault.address, MAX_UINT256);
+  });
+
+  enum LinearPoolState {
+    BALANCED,
+    MAIN_EXCESS,
+    MAIN_LACK,
+  }
+
+  function itRebalancesThePool(expectedState: LinearPoolState) {
+    it('rebalance the pool', async () => {
+      const { lowerTarget, upperTarget } = await pool.getTargets();
+
+      const { cash } = await vault.getPoolTokenInfo(poolId, BRZ);
+      const scaledCash = cash.mul(BRZ_SCALING);
+
+      let fees;
+      if (scaledCash.gt(upperTarget)) {
+        expect(expectedState).to.equal(LinearPoolState.MAIN_EXCESS);
+
+        const excess = scaledCash.sub(upperTarget);
+        fees = excess.mul(SWAP_FEE_PERCENTAGE).div(FP_ONE);
+      } else if (scaledCash.lt(lowerTarget)) {
+        expect(expectedState).to.equal(LinearPoolState.MAIN_LACK);
+
+        const lack = lowerTarget.sub(scaledCash);
+        fees = lack.mul(SWAP_FEE_PERCENTAGE).div(FP_ONE);
+      } else {
+        expect(expectedState).to.equal(LinearPoolState.BALANCED);
+
+        fees = 0;
+      }
+
+      const initialRecipientMainBalance = await brz.balanceOf(other.address);
+
+      if (expectedState != LinearPoolState.BALANCED) {
+        await rebalancer.connect(holder).rebalance(other.address);
+      } else {
+        await rebalancer.connect(holder).rebalanceWithExtraMain(other.address, 5);
+      }
+
+      const finalRecipientMainBalance = await brz.balanceOf(other.address);
+
+      if (fees > 0) {
+        // The recipient of the rebalance call should get the fees that were collected (though there's some rounding
+        // error in the main-wrapped conversion).
+        expect(finalRecipientMainBalance.sub(initialRecipientMainBalance)).to.be.almostEqual(
+          fees.div(BRZ_SCALING),
+          0.000001 // old: 0.00000001
+        );
+      } else {
+        // The recipient of the rebalance call will get any extra main tokens that were not utilized.
+        expect(finalRecipientMainBalance).to.be.almostEqual(initialRecipientMainBalance, 0.00000001);
+      }
+
+      const mainInfo = await vault.getPoolTokenInfo(poolId, BRZ);
+
+      const expectedMainBalance = lowerTarget.add(upperTarget).div(2);
+      expect(mainInfo.cash.mul(BRZ_SCALING)).to.equal(expectedMainBalance);
+      expect(mainInfo.managed).to.equal(0);
+    });
+  }
+
+  describe('create and check getters', () => {
+    it('deploy a linear pool', async () => {
+      const tx = await factory.create(
+        '',
+        '',
+        BRZ,
+        cBRZ,
+        INITIAL_UPPER_TARGET,
+        SWAP_FEE_PERCENTAGE,
+        owner.address,
+        PROTOCOL_ID
+      );
+      const event = expectEvent.inReceipt(await tx.wait(), 'PoolCreated');
+
+      pool = await task.instanceAt('MidasLinearPool', event.args.pool);
+      expect(await factory.isPoolFromFactory(pool.address)).to.be.true;
+
+      poolId = await pool.getPoolId();
+      const [registeredAddress] = await vault.getPool(poolId);
+      expect(registeredAddress).to.equal(pool.address);
+
+      const { assetManager } = await vault.getPoolTokenInfo(poolId, BRZ); // We could query for either BRZ or cBRZ
+      rebalancer = await task.instanceAt('MidasLinearPoolRebalancer', assetManager);
+
+      await brz.connect(holder).approve(rebalancer.address, MAX_UINT256); // To send extra main on rebalance
+    });
+
+    it('check factory version', async () => {
+      const expectedFactoryVersion = {
+        name: 'MidasLinearPoolFactory',
+        version: 1,
+        deployment: '20230310-midas-rebalanced-linear-pool',
+      };
+
+      expect(await factory.version()).to.equal(JSON.stringify(expectedFactoryVersion));
+    });
+
+    it('check pool version', async () => {
+      const expectedPoolVersion = {
+        name: 'MidasLinearPool',
+        version: 1,
+        deployment: '20230310-midas-rebalanced-linear-pool',
+      };
+
+      expect(await pool.version()).to.equal(JSON.stringify(expectedPoolVersion));
+    });
+  });
+
+  describe('join, and rebalance', () => {
+    it('join the pool', async () => {
+      // We're going to join with enough main token to bring the Pool above its upper target, which will let us later
+      // rebalance.
+
+      const joinAmount = INITIAL_UPPER_TARGET.mul(2).div(BRZ_SCALING);
+
+      await vault
+        .connect(holder)
+        .swap(
+          { kind: SwapKind.GivenIn, poolId, assetIn: BRZ, assetOut: pool.address, amount: joinAmount, userData: '0x' },
+          { sender: holder.address, recipient: holder.address, fromInternalBalance: false, toInternalBalance: false },
+          0,
+          MAX_UINT256
+        );
+
+      // Assert join amount - some fees will be collected as we're going over the upper target.
+      const excess = joinAmount.mul(BRZ_SCALING).sub(INITIAL_UPPER_TARGET);
+      const joinCollectedFees = excess.mul(SWAP_FEE_PERCENTAGE).div(FP_ONE);
+
+      const expectedBPT = joinAmount.mul(BRZ_SCALING).sub(joinCollectedFees);
+      expect(await pool.balanceOf(holder.address)).to.equal(expectedBPT);
+    });
+
+    itRebalancesThePool(LinearPoolState.MAIN_EXCESS);
+
+    it('set final targets', async () => {
+      await pool.connect(owner).setTargets(FINAL_LOWER_TARGET, FINAL_UPPER_TARGET);
+    });
+  });
+
+  describe('generate excess of main token and rebalance', () => {
+    it('deposit main tokens', async () => {
+      // We're going to join with enough main token to bring the Pool above its upper target, which will let us later
+      // rebalance.
+
+      const { upperTarget } = await pool.getTargets();
+      const joinAmount = upperTarget.mul(5).div(BRZ_SCALING);
+
+      await vault
+        .connect(holder)
+        .swap(
+          { kind: SwapKind.GivenIn, poolId, assetIn: BRZ, assetOut: pool.address, amount: joinAmount, userData: '0x' },
+          { sender: holder.address, recipient: holder.address, fromInternalBalance: false, toInternalBalance: false },
+          0,
+          MAX_UINT256
+        );
+    });
+
+    itRebalancesThePool(LinearPoolState.MAIN_EXCESS);
+  });
+
+  describe('generate lack of main token and rebalance', () => {
+    it('withdraw main tokens', async () => {
+      // We're going to withdraw enough man token to bring the Pool below its lower target, which will let us later
+      // rebalance.
+
+      const { cash } = await vault.getPoolTokenInfo(poolId, BRZ);
+      const scaledCash = cash.mul(BRZ_SCALING);
+      const { lowerTarget } = await pool.getTargets();
+
+      const exitAmount = scaledCash.sub(lowerTarget.div(3)).div(BRZ_SCALING);
+
+      await vault.connect(holder).swap(
+        {
+          kind: SwapKind.GivenOut,
+          poolId,
+          assetIn: pool.address,
+          assetOut: BRZ,
+          amount: exitAmount,
+          userData: '0x',
+        },
+        { sender: holder.address, recipient: holder.address, fromInternalBalance: false, toInternalBalance: false },
+        MAX_UINT256,
+        MAX_UINT256
+      );
+    });
+
+    itRebalancesThePool(LinearPoolState.MAIN_LACK);
+  });
+
+  describe('join below upper target and rebalance', () => {
+    it('deposit main tokens', async () => {
+      // We're going to join with few tokens, causing the Pool to not reach its upper target.
+
+      const { lowerTarget, upperTarget } = await pool.getTargets();
+      const midpoint = lowerTarget.add(upperTarget).div(2);
+
+      const joinAmount = midpoint.div(100).div(BRZ_SCALING);
+
+      await vault
+        .connect(holder)
+        .swap(
+          { kind: SwapKind.GivenIn, poolId, assetIn: BRZ, assetOut: pool.address, amount: joinAmount, userData: '0x' },
+          { sender: holder.address, recipient: holder.address, fromInternalBalance: false, toInternalBalance: false },
+          0,
+          MAX_UINT256
+        );
+    });
+
+    itRebalancesThePool(LinearPoolState.BALANCED);
+  });
+
+  describe('exit above lower target and rebalance', () => {
+    it('withdraw main tokens', async () => {
+      // We're going to exit with few tokens, causing for the Pool to not reach its lower target.
+
+      const { lowerTarget, upperTarget } = await pool.getTargets();
+      const midpoint = lowerTarget.add(upperTarget).div(2);
+
+      const exitAmount = midpoint.div(100).div(BRZ_SCALING);
+
+      await vault.connect(holder).swap(
+        {
+          kind: SwapKind.GivenOut,
+          poolId,
+          assetIn: pool.address,
+          assetOut: BRZ,
+          amount: exitAmount,
+          userData: '0x',
+        },
+        { sender: holder.address, recipient: holder.address, fromInternalBalance: false, toInternalBalance: false },
+        MAX_UINT256,
+        MAX_UINT256
+      );
+    });
+
+    itRebalancesThePool(LinearPoolState.BALANCED);
+  });
+
+  describe('rebalance repeatedly', () => {
+    itRebalancesThePool(LinearPoolState.BALANCED);
+    itRebalancesThePool(LinearPoolState.BALANCED);
+  });
+
+  describe('rebalancer query protection', () => {
+    it('reverts with a malicious lending pool', async () => {
+      const { cash } = await vault.getPoolTokenInfo(poolId, BRZ);
+      const scaledCash = cash.mul(BRZ_SCALING);
+      const { lowerTarget } = await pool.getTargets();
+
+      const exitAmount = scaledCash.sub(lowerTarget.div(3)).div(BRZ_SCALING);
+
+      await vault.connect(holder).swap(
+        {
+          kind: SwapKind.GivenOut,
+          poolId,
+          assetIn: pool.address,
+          assetOut: BRZ,
+          amount: exitAmount,
+          userData: '0x',
+        },
+        { sender: holder.address, recipient: holder.address, fromInternalBalance: false, toInternalBalance: false },
+        MAX_UINT256,
+        MAX_UINT256
+      );
+
+      await setCode(cBRZ, getExternalPackageArtifact('linear-pools/MockCToken').deployedBytecode);
+      const mockMaliciousMidasToken = await getExternalPackageDeployedAt('linear-pools/MockCToken', cBRZ);
+
+      await mockMaliciousMidasToken.setRevertType(2); // Type 2 is malicious swap query revert
+      await expect(rebalancer.rebalance(other.address)).to.be.revertedWith('BAL#357'); // MALICIOUS_QUERY_REVERT
+    });
+  });
+});
