--- pkg/linear-pools/contracts/midas-linear-pool/README.md
@@ -1,2 +1,0 @@
-| :warning: | The Midas Linear Pool is still under development and should not be deployed to production networks. It may contain bugs and/or change significantly. |
-| --------- | :--------------------------------------------------------------------------------------------------------------------------------------------------- |
